requirements.txt
google-genai>=0.2.0 